requirements.txt
fastapi==0.93.0
python-multipart>=0.0.6
uvicorn[standard]==0.20.0
dependency-injector==4.49.0
pytest==7.2.2
httpx==0.23.3
pydantic==1.10.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator>=2.1.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1